Add controls functions in scene
