+++ src/Game/TestScene.cpp
@@ -31,6 +31,25 @@ void TestScene::addGameObject(std::string luaScript) {
 	game_objects_.emplace_back(gaf.create(luaScript));
 }
 
+void TestScene::key_controls(const bool* keys, float delta_time) {
+	auto& renderer = ReflexEngine::get_instance().renderer_;
+	auto& camera = ReflexEngine::get_instance().camera_;
+
+	if (keys[GLFW_KEY_W]) camera.move(CameraMovement::forward, delta_time);
+	if (keys[GLFW_KEY_S]) camera.move(CameraMovement::backward, delta_time);
+	if (keys[GLFW_KEY_A]) camera.move(CameraMovement::left, delta_time);
+	if (keys[GLFW_KEY_D]) camera.move(CameraMovement::right, delta_time);
+
+	if (keys[GLFW_KEY_C]) renderer.toggle_wireframe();
+	if (keys[GLFW_KEY_Z]) camera.toggle_noclip();
+}
+
+void TestScene::mouse_controls(float xpos, float ypos) {
+	auto& camera = ReflexEngine::get_instance().camera_;
+
+	camera.mouse_move(xpos, ypos);
+}
+
 void TestScene::add_draw_call() {
 	auto& renderer = ReflexEngine::get_instance().renderer_;
 
@@ -45,10 +64,10 @@ void TestScene::add_draw_call() {
 void TestScene::update(float delta_time) {
 	const auto& camera = ReflexEngine::get_instance().camera_;
 
-	glm::vec3 lower_light = camera.GetCamPosition();
+	glm::vec3 lower_light = camera.get_position();
 	lower_light.y -= 0.3f;
 
-	glm::vec3 cam_direction = camera.GetCamDirection();
+	glm::vec3 cam_direction = camera.get_direction();
 
 	flashlight_.SetFlash(lower_light, cam_direction);
